# specs/001-fullstack-web-app/contracts/openapi.yaml
openapi: 3.0.0
info:
  title: Todo Application API
  description: API for the full-stack todo application with JWT authentication
  version: 1.0.0
servers:
  - url: http://localhost:8000/api
    description: Local development server

paths:
  /auth/token:
    post:
      summary: Create JWT token
      description: Authenticate user and return JWT token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                password:
                  type: string
                  format: password
              required:
                - email
                - password
      responses:
        '200':
          description: Successful authentication
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                  user:
                    $ref: '#/components/schemas/User'
        '401':
          description: Invalid credentials

  /todos:
    get:
      summary: Get user's todos
      description: Retrieve all todos for the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of user's todos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
    post:
      summary: Create a new todo
      description: Create a new todo for the authenticated user
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoCreate'
      responses:
        '201':
          description: Todo created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '400':
          description: Invalid input

  /todos/{id}:
    get:
      summary: Get a specific todo
      description: Retrieve a specific todo by ID for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Todo details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
    put:
      summary: Update a todo
      description: Update an existing todo for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TodoUpdate'
      responses:
        '200':
          description: Todo updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
        '404':
          description: Todo not found
    delete:
      summary: Delete a todo
      description: Delete an existing todo for the authenticated user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Todo deleted successfully
        '404':
          description: Todo not found

  /users/me:
    get:
      summary: Get current user profile
      description: Retrieve the profile of the authenticated user
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    User:
      type: object
      properties:
        user_id:
          type: string
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        first_name:
          type: string
        last_name:
          type: string
        role:
          type: string
          enum: [user, admin, moderator]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_active:
          type: boolean
      required:
        - user_id
        - email
        - email_verified
        - role
        - created_at
        - updated_at
        - is_active

    Todo:
      type: object
      properties:
        todo_id:
          type: string
        user_id:
          type: string
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        completed:
          type: boolean
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
      required:
        - todo_id
        - user_id
        - title
        - completed
        - created_at
        - updated_at

    TodoCreate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        completed:
          type: boolean
          default: false
      required:
        - title

    TodoUpdate:
      type: object
      properties:
        title:
          type: string
          maxLength: 200
        description:
          type: string
          maxLength: 1000
        completed:
          type: boolean